backend/pyproject.toml
[project]
name = "ai-candidate-evaluator"
version = "0.1.0"
description = "Email-driven AI candidate screening agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3.2",
    "anthropic>=0.39",
    "google-api-python-client>=2.150",
    "google-auth>=2.35",
    "google-auth-oauthlib>=1.2",
    "httpx>=0.27",
    "pymupdf>=1.24",
    "python-docx>=1.1",
    "beautifulsoup4>=4.12",
    "structlog>=24.4",
    "python-multipart>=0.0.12",
    "tenacity>=9.0",
    "python-jose[cryptography]>=3.3",
    "playwright>=1.47",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
markers = [
    "live: tests that hit real external services (Gmail, Anthropic, GitHub) — opt in with `pytest -m live` or RUN_LIVE_TESTS=1",
]
# By default, exclude live tests from the regular pytest run. CI and local
# devs need to opt in explicitly so we never accidentally hammer real APIs.
addopts = "-m 'not live'"
